Internal: Restored initial image formatting, added ImageResize plugin dedicated for the resizing.
