+++ src/image/imageediting.js
@@ -59,25 +59,7 @@ export default class ImageEditing extends Plugin {
 
 		conversion.for( 'editingDowncast' ).elementToElement( {
 			model: 'image',
-			view: ( modelElement, viewWriter ) => {
-				const ret = toImageWidget( createImageViewElement( viewWriter ), viewWriter, t( 'image widget' ) );
-
-				editor.plugins.get( 'WidgetResizer' ).apply( ret, viewWriter, {
-					getResizeHost( wrapper ) {
-						return wrapper.querySelector( 'img' );
-					},
-					getAspectRatio( resizeHost ) {
-						return resizeHost.naturalWidth / resizeHost.naturalHeight;
-					},
-					isCentered( context ) {
-						const imageStyle = context._getModel( editor, context.widgetWrapperElement ).getAttribute( 'imageStyle' );
-
-						return !imageStyle || imageStyle == 'full';
-					}
-				} );
-
-				return ret;
-			}
+			view: ( modelElement, viewWriter ) => toImageWidget( createImageViewElement( viewWriter ), viewWriter, t( 'image widget' ) )
 		} );
 
 		conversion.for( 'downcast' )
