+++ tests/manual/image.html
@@ -1,7 +1,7 @@
 <div id="editor">
 	<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla finibus consequat placerat. Vestibulum id tellus et mauris sagittis tincidunt quis id mauris. Curabitur consectetur lectus sit amet tellus mattis, non lobortis leo interdum. </p>
 	<figure class="image image-style-side">
-		<img src="sample.jpg" alt="" height="60" />
+		<img src="sample.jpg" alt="" />
 	</figure>
 	<figure class="image">
 		<img src="sample.jpg" alt="" />
